refactore: add utils module
